Layout and Fixes
- Fixed years worked calculation.

- Removed the auto enddate selection.

NOTE:
---------------------------------

Holiday handling:

In order to correctly represent the time taken when a vacation is split
by a holiday, the user will need to add multiple entries for a single
vacation. Ex.(If someone was taking a vacation on the week of July 4th.
You need to add one entry for 7/2 to 7/3 (16 hours) and another entry
for 7/5 to 7/9 (24 hours)).

----------------------------------

-A little more layout compression.

-Disabled the Accruals run button.

-Fixed some printing typos