Vacation Report Bugfix
WhoOnVaca sub wasn't requesting it's results to be ordered by date and
was outputting gibberish. I have no idea how it ever worked before like
this. My only guess is that the previous version of the MySQL server
defaulted to that particualar order by. Now it appears not to. Added
order by.  :-o